requirements-dev.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
-e .
django==2.0.2             # via drf-jwt-2fa
djangorestframework==3.7.7  # via drf-jwt-2fa
djangorestframework-jwt==1.11.0  # via drf-jwt-2fa
pluggy==0.6.0             # via tox
py==1.5.2                 # via tox
pyjwt==1.5.3              # via djangorestframework-jwt
pytz==2017.3              # via django
six==1.11.0               # via tox
tox==2.9.1
virtualenv==15.1.0        # via tox
